Fixes tests for CRAN

+++ tests/testthat/test-file_caching.R
@@ -1,15 +1,18 @@
 
 context("Cache directory handling")
 
+# set up the cache in tempdir() for testing on CRAN
+manage_cache$cache_path_set("foobar", type = "tempdir")
+
 # test that .set_cache() creates a cache directory if none exists --------------
 test_that("test that set_cache creates a cache directory if none exists", {
-  skip_on_cran()
-  unlink(rappdirs::user_cache_dir("bomrang"), recursive = TRUE)
   cache <- TRUE
   .set_cache(cache)
   expect_true(dir.exists(manage_cache$cache_path_get()))
   # cleanup
-  unlink(manage_cache$cache_path_get())
+  unlink(manage_cache$cache_path_get(),
+         recursive = TRUE,
+         force = TRUE)
 })
 
 # test that .set_cache() does a cache directory if cache is FALSE --------------
@@ -23,8 +26,12 @@ test_that("test that set_cache does not create a dir if cache == FALSE", {
 
 test_that("cache directory is created if necessary", {
   # if cache directory exists during testing, remove it
-  unlink(manage_cache$cache_path_get(),
-         recursive = TRUE)
+  # clean up before testing
+  if (dir.exists(manage_cache$cache_path_get())) {
+    unlink(manage_cache$cache_path_get(),
+           recursive = TRUE,
+           force = TRUE)
+  }
   cache <- TRUE
   cache_dir <- .set_cache(cache)
   expect_true(dir.exists(manage_cache$cache_path_get()))
@@ -34,25 +41,24 @@ test_that("cache directory is created if necessary", {
 
 test_that("caching utils list files in cache and delete when asked", {
   skip_on_cran()
-  unlink(manage_cache$cache_path_get())
   f <-
     raster::raster(system.file("external/test.grd", package = "raster"))
   cache_dir <- manage_cache$cache_path_get()
   raster::writeRaster(f, file.path(cache_dir, "file1.tif"), format = "GTiff")
   raster::writeRaster(f, file.path(cache_dir, "file2.tif"), format = "GTiff")
-
+  
   # test bomrang cache list
   k <- basename(manage_cache$list())
   expect_equal(basename(manage_cache$list()), k)
-
+  
   # file should not exist, expect error
   expect_error(manage_cache$delete("file1.asc"))
-
+  
   # test delete one file
   manage_cache$delete("file1.tif")
   l <- basename(manage_cache$list())
   expect_equal(basename(manage_cache$list()), l)
-
+  
   # test delete all
   manage_cache$delete_all()
   expect_equal(manage_cache$list(),
